KILL THE BUG!
Got the little dalek. After one of Stacks update snp loci are recorded
only as loci names, w/o BP separated by _, so the gsub way to extract
them should be gsub("_." instead of gsub("_._.|_.._." as was before.
This was causing the loci count to be erroneous. Happy morning chasing
this.